pyproject.toml
[project]
name = "greptime-pymysql-example"
version = "0.1.0"
description = "An example of connecting to GreptimeCloud using PyMySQL."
authors = [
    { name = "nervouna", email = "xiaoyuguan@hotmail.com" }
]
dependencies = [
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.0",
    "psutil>=5.9.5",
]
readme = "README.md"
requires-python = ">= 3.8"
license = { text = "WTFPL" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.rye.scripts]
up = "python main.py"

[tool.hatch.metadata]
allow-direct-references = true
